+++ Mile.Helpers/Mile.Helpers.h
@@ -124,4 +124,18 @@ EXTERN_C HRESULT WINAPI MileGetWindowSystemBackdropTypeAttribute(
     _In_ HWND WindowHandle,
     _Out_ PMILE_WINDOW_SYSTEM_BACKDROP_TYPE Value);
 
+/**
+ * @brief Specifies the system-drawn backdrop material of a window, including
+ *        behind the non-client area.
+ * @param WindowHandle The handle to the window for which the attribute value
+ *                     is to be set.
+ * @param Value Flags for specifying the system-drawn backdrop material of a
+ *              window, including behind the non-client area.
+ * @return If the function succeeds, it returns S_OK. Otherwise, it returns an
+ *         HRESULT error code.
+*/
+EXTERN_C HRESULT WINAPI MileSetWindowSystemBackdropTypeAttribute(
+    _In_ HWND WindowHandle,
+    _In_ MILE_WINDOW_SYSTEM_BACKDROP_TYPE Value);
+
 #endif // !MILE_WINDOWS_HELPERS