+++ ReleaseNotes.md
@@ -10,3 +10,4 @@
 - Add MileSetWindowUseImmersiveDarkModeAttribute function.
 - Add MILE_WINDOW_SYSTEM_BACKDROP_TYPE enum.
 - Add MileGetWindowSystemBackdropTypeAttribute function.
+- Add MileSetWindowSystemBackdropTypeAttribute function.